refactor(issues): split the 1.6k-line index.ts into schema + registry
index.ts mixed all 20 spread-type definitions with the archive, so every issue touched a 1,632-line file (the merge-conflict magnet PUBLISHING.md §VII has a protocol for). Moved the type layer to schema.ts (1,483 lines); index.ts is now a 169-line registry (issue imports + ALL_ISSUES + LATEST_ISSUE + findIssue) re-exporting the schema via 'export * from ./schema', so every 'from ./issues' / 'from ./index' import resolves unchanged.

Verified: tsc clean, full vite build clean (77 modules), editorial gate + accent tests green.
